+++ pdk-theme-options/pdk-theme-options-plugin.php
@@ -3,7 +3,7 @@
  * Plugin Name: PDK Theme Options
  * Plugin URI:  https://github.com/PDK-Online-Succes/pdk-theme-options-plugin
  * Description: Centrale beheerplugin voor PDK Online Succes klanten. Modulaire WordPress/WooCommerce functionaliteiten in één plugin.
- * Version:     2.8.0
+ * Version:     2.8.1
  * Author:      PDK Online Succes
  * Author URI:  https://pdk.nl
  * License:     GPL-2.0-or-later
@@ -21,7 +21,7 @@ defined( 'ABSPATH' ) || exit;
 
 
 // Plugin constants.
-define( 'PDK_PLUGIN_VERSION', '2.8.0' );
+define( 'PDK_PLUGIN_VERSION', '2.8.1' );
 define( 'PDK_PLUGIN_FILE',    __FILE__ );
 define( 'PDK_PLUGIN_DIR',     plugin_dir_path( __FILE__ ) );
 define( 'PDK_PLUGIN_URL',     plugin_dir_url( __FILE__ ) );